# PERF: Re-use the faster get_open_and_closes from tradingcalendar