Updated tradingcalendar_bmf and _tse to use the faster get_open_and_closes from tradingcalendar.

Before:
```
In [1]: %time import zipline.utils.tradingcalendar
CPU times: user 1.48 s, sys: 139 ms, total: 1.61 s
Wall time: 1.73 s

In [2]: %time import zipline.utils.tradingcalendar_bmf
CPU times: user 12.7 s, sys: 108 ms, total: 12.8 s
Wall time: 13.1 s

In [3]: %time import zipline.utils.tradingcalendar_tse
CPU times: user 12.1 s, sys: 108 ms, total: 12.2 s
Wall time: 12.4 s
```
After:
```
In [1]: %time import zipline.utils.tradingcalendar
CPU times: user 1.27 s, sys: 103 ms, total: 1.38 s
Wall time: 1.44 s

In [2]: %time import zipline.utils.tradingcalendar_bmf
CPU times: user 1.31 s, sys: 26.4 ms, total: 1.34 s
Wall time: 1.65 s

In [3]: %time import zipline.utils.tradingcalendar_tse
CPU times: user 773 ms, sys: 7.93 ms, total: 781 ms
Wall time: 782 ms
```