@ssanderson 2015-09-08
@richafrank looks like this missed `tradingcalendar_lse.py`?  Though that module doesn't appear to have an `open_and_closes`, so I'm not sure if it works at all right now.