@ssanderson 2015-09-08
```
In [1]: from pandas import *

In [2]: from zipline.utils.tradingcalendar_tse import open_and_closes as tse

In [3]: old_tse = read_pickle('tse.pkl')

In [4]: (old_tse != tse).sum()
Out[4]:
market_open     0
market_close    0
dtype: int64

In [5]: from zipline.utils.tradingcalendar_bmf import open_and_closes as bmf

In [6]: old_bmf = read_pickle('bmf.pkl')

In [7]: (old_bmf != bmf).sum()
Out[7]:
market_open     0
market_close    0
dtype: int64
```