@ssanderson 2015-09-08
LGTM.  This would probably be even faster if we we're doing a `map`, but it's better than before.