@richafrank 2015-09-08
Awesome - thanks for the pickle check.  Yea, since tradingcalendar_lse.py didn't have ```open_and_closes```, I didn't touch it.