@ssanderson 2015-09-08
@richafrank FWIW, doing this vectorized is about 40% faster than what we're currently doing:

```
def get_open_and_closes(trading_days,
                        early_closes,
                        open_delta,
                        close_delta,
                        early_close_delta):
    # Opens are always at 9:30 US/Eastern.
    o = (cast_tz(trading_days, 'US/Eastern') + open_delta).tz_convert('UTC')
    c = cast_tz(trading_days, 'US/Eastern').values

    early_close_indices = trading_days.get_indexer(early_closes)
    regular_close_indices = np.setdiff1d(
        np.arange(len(trading_days)),
        early_close_indices,
    )

    # Add early_close_delta to early closes, and close_delta to regular_closes.
    c[early_close_indices] += early_close_delta.asm8
    c[regular_close_indices] += close_delta.asm8

    return pd.DataFrame(
        data={'market_open': o, 'market_close': pd.DatetimeIndex(c, tz='UTC')},
        # Explicitly force this column order.
        columns=['market_open', 'market_close'],
        index=trading_days,
    )


open_and_closes = get_open_and_closes(
    trading_days,
    early_closes,
    open_delta=pd.Timedelta('9 hours 31 minutes'),
    close_delta=pd.Timedelta('16 hours'),
    early_close_delta=pd.Timedelta('13 hours'),
)
```

**Old Version:**
```
In [1]: %time from zipline.utils.tradingcalendar import open_and_closes
CPU times: user 823 ms, sys: 31.7 ms, total: 854 ms
Wall time: 854 ms
```


**New Version:**
```
In [1]: %time from zipline.utils.tradingcalendar import open_and_closes
CPU times: user 468 ms, sys: 47.1 ms, total: 515 ms
Wall time: 514 ms
```